+++ llvm/include/llvm/Support/Compression.h
@@ -42,8 +42,6 @@ Error uncompress(StringRef InputBuffer,
                  SmallVectorImpl<char> &UncompressedBuffer,
                  size_t UncompressedSize);
 
-uint32_t crc32(StringRef Buffer);
-
 } // End of namespace zlib
 
 namespace profile = llvm::compression::zlib;